Use uint256 arguments in push: safe and saves gas.

+++ lot-ethereum/lottery1.sol
@@ -70,14 +70,14 @@ contract OrchidLottery1 {
     }
 
 
-    function push(address signer, uint128 recover, uint128 transfer) external payable {
+    function push(address signer, uint256 recover, uint256 transfer) external payable {
         Pot storage pot = find(msg.sender, signer);
 
         uint256 escrow = pot.escrow_;
         uint256 amount = pot.amount_ + msg.value;
 
         if (recover != 0) {
-            uint128 warned = pot.warned_;
+            uint256 warned = pot.warned_;
             require(pot.unlock_ - 1 < block.timestamp);
             require(recover <= warned);
             amount += recover;
@@ -85,7 +85,7 @@ contract OrchidLottery1 {
             warned -= recover;
             if (warned == 0)
                 pot.unlock_ = 0;
-            pot.warned_ = warned;
+            pot.warned_ = uint128(warned);
         }
 
         require(transfer <= amount);
